remove RethinkDB.TeamMember
