+++ codegen.json
@@ -92,7 +92,7 @@
           "NotifyTaskInvolves": "../../database/types/NotificationTaskInvolves#default",
           "NotifyTeamArchived": "../../database/types/NotificationTeamArchived#default",
           "Organization": "./types/Organization#OrganizationSource",
-          "OrganizationUser": "../../postgres/types/index#OrganizationUser",
+          "OrganizationUser": "../../postgres/types/index#OrganizationUser as OrganizationUserDB",
           "PokerMeeting": "../../database/types/MeetingPoker#default as MeetingPoker",
           "PokerMeetingMember": "../../database/types/MeetingPokerMeetingMember#default as PokerMeetingMemberDB",
           "PokerTemplate": "../../database/types/PokerTemplate#default as PokerTemplateDB",
@@ -125,7 +125,7 @@
           "TeamHealthPhase": "./types/TeamHealthPhase#TeamHealthPhaseSource",
           "TeamHealthStage": "./types/TeamHealthStage#TeamHealthStageSource",
           "TeamInvitation": "../../database/types/TeamInvitation#default",
-          "TeamMember": "../../database/types/TeamMember#default as TeamMemberDB",
+          "TeamMember": "../../postgres/types/index#TeamMember as TeamMember",
           "TeamMemberIntegrationAuthOAuth2": "../../postgres/queries/getTeamMemberIntegrationAuth#TeamMemberIntegrationAuth",
           "TeamMemberIntegrations": "./types/TeamMemberIntegrations#TeamMemberIntegrationsSource",
           "TeamPromptMeeting": "../../database/types/MeetingTeamPrompt#default as MeetingTeamPromptDB",
