+++ packages/server/dataloader/foreignKeyLoaderMakers.ts
@@ -6,6 +6,23 @@ export const teamsByOrgIds = foreignKeyLoaderMaker('teams', 'orgId', (orgIds) =>
   selectTeams().where('orgId', 'in', orgIds).where('isArchived', '=', false).execute()
 )
 
+export const teamMembersByTeamId = foreignKeyLoaderMaker('teamMembers', 'teamId', (teamIds) =>
+  getKysely()
+    .selectFrom('TeamMember')
+    .selectAll()
+    .where('teamId', 'in', teamIds)
+    .where('isNotRemoved', '=', true)
+    .execute()
+)
+
+export const teamMembersByUserId = foreignKeyLoaderMaker('teamMembers', 'userId', (userIds) =>
+  getKysely()
+    .selectFrom('TeamMember')
+    .selectAll()
+    .where('userId', 'in', userIds)
+    .where('isNotRemoved', '=', true)
+    .execute()
+)
 export const discussionsByMeetingId = foreignKeyLoaderMaker(
   'discussions',
   'meetingId',
