+++ packages/server/dataloader/rethinkForeignKeyLoaderMakers.ts
@@ -221,31 +221,3 @@ export const teamInvitationsByTeamId = new RethinkForeignKeyLoaderMaker(
       .run()
   }
 )
-
-export const teamMembersByTeamId = new RethinkForeignKeyLoaderMaker(
-  'teamMembers',
-  'teamId',
-  async (teamIds) => {
-    // tasksByUserId is expensive since we have to look up each team to check the team archive status
-    const r = await getRethink()
-    return r
-      .table('TeamMember')
-      .getAll(r.args(teamIds), {index: 'teamId'})
-      .filter({isNotRemoved: true})
-      .run()
-  }
-)
-
-export const teamMembersByUserId = new RethinkForeignKeyLoaderMaker(
-  'teamMembers',
-  'userId',
-  async (userIds) => {
-    // tasksByUserId is expensive since we have to look up each team to check the team archive status
-    const r = await getRethink()
-    return r
-      .table('TeamMember')
-      .getAll(r.args(userIds), {index: 'userId'})
-      .filter({isNotRemoved: true})
-      .run()
-  }
-)
