+++ packages/server/dataloader/rethinkPrimaryKeyLoaderMakers.ts
@@ -18,6 +18,5 @@ export const slackAuths = new RethinkPrimaryKeyLoaderMaker('SlackAuth')
 export const slackNotifications = new RethinkPrimaryKeyLoaderMaker('SlackNotification')
 export const suggestedActions = new RethinkPrimaryKeyLoaderMaker('SuggestedAction')
 export const tasks = new RethinkPrimaryKeyLoaderMaker('Task')
-export const teamMembers = new RethinkPrimaryKeyLoaderMaker('TeamMember')
 export const teamInvitations = new RethinkPrimaryKeyLoaderMaker('TeamInvitation')
 export const templateDimensions = new RethinkPrimaryKeyLoaderMaker('TemplateDimension')
