+++ packages/server/graphql/public/mutations/uploadUserImage.ts
@@ -1,21 +1,18 @@
 import {SubscriptionChannel} from 'parabol-client/types/constEnums'
-import getRethink from '../../../database/rethinkDriver'
 import getFileStoreManager from '../../../fileStorage/getFileStoreManager'
 import normalizeAvatarUpload from '../../../fileStorage/normalizeAvatarUpload'
 import validateAvatarUpload from '../../../fileStorage/validateAvatarUpload'
 import getKysely from '../../../postgres/getKysely'
-import updateUser from '../../../postgres/queries/updateUser'
 import {getUserId, isAuthenticated} from '../../../utils/authorization'
 import publish from '../../../utils/publish'
 import standardError from '../../../utils/standardError'
-import {MutationResolvers, TeamMember} from '../resolverTypes'
+import {MutationResolvers} from '../resolverTypes'
 
 const uploadUserImage: MutationResolvers['uploadUserImage'] = async (
   _,
   {file},
   {authToken, dataLoader, socketId: mutatorId}
 ) => {
-  const r = await getRethink()
   const pg = getKysely()
   const operationId = dataLoader.share()
   const subOptions = {operationId, mutatorId}
@@ -34,24 +31,14 @@ const uploadUserImage: MutationResolvers['uploadUserImage'] = async (
   const manager = getFileStoreManager()
   const publicLocation = await manager.putUserAvatar(normalBuffer, userId, normalExt)
 
-  await Promise.all([
-    pg
-      .with('TeamMemberUpdate', (qc) =>
-        qc.updateTable('TeamMember').set({picture: publicLocation}).where('userId', '=', userId)
-      )
-      .updateTable('User')
-      .set({picture: publicLocation})
-      .where('id', '=', userId)
-      .execute(),
-    r
-      .table('TeamMember')
-      .getAll(userId, {index: 'userId'})
-      .update({picture: publicLocation}, {returnChanges: true})('changes')('new_val')
-      .default([])
-      .run() as unknown as TeamMember[],
-
-    updateUser({picture: publicLocation}, userId)
-  ])
+  await pg
+    .with('TeamMemberUpdate', (qc) =>
+      qc.updateTable('TeamMember').set({picture: publicLocation}).where('userId', '=', userId)
+    )
+    .updateTable('User')
+    .set({picture: publicLocation})
+    .where('id', '=', userId)
+    .execute()
   dataLoader.clearAll(['users', 'teamMembers'])
   const teamMembers = await dataLoader.get('teamMembersByUserId').load(userId)
   const teamIds = teamMembers.map(({teamId}) => teamId)
