+++ packages/server/graphql/public/typeDefs/RepoIntegrationQueryPayload.graphql
@@ -1,0 +1,16 @@
+"""
+The details associated with the possible repo and project integrations
+"""
+type RepoIntegrationQueryPayload {
+  error: StandardMutationError
+
+  """
+  true if the items returned are a subset of all the possible integration, else false (all possible integrations)
+  """
+  hasMore: Boolean
+
+  """
+  All the integrations that are likely to be integrated
+  """
+  items: [RepoIntegration!]
+}
