+++ packages/server/graphql/public/types/AgendaItem.ts
@@ -3,7 +3,7 @@ import {AgendaItemResolvers} from '../resolverTypes'
 const AgendaItem: AgendaItemResolvers = {
   isActive: ({isActive}) => !!isActive,
   teamMember: async ({teamMemberId}, _args: unknown, {dataLoader}) => {
-    return dataLoader.get('teamMembers').load(teamMemberId)
+    return dataLoader.get('teamMembers').loadNonNull(teamMemberId)
   }
 }
