+++ packages/server/graphql/resolvers.ts
@@ -6,9 +6,9 @@ import GenericMeetingStage from '../database/types/GenericMeetingStage'
 import Meeting from '../database/types/Meeting'
 import Organization from '../database/types/Organization'
 import Task from '../database/types/Task'
-import TeamMember from '../database/types/TeamMember'
 import User from '../database/types/User'
 import {Loaders} from '../dataloader/RootDataLoader'
+import {TeamMember} from '../postgres/types'
 import {AnyMeeting} from '../postgres/types/Meeting'
 import {getUserId, isSuperUser, isUserBillingLeader} from '../utils/authorization'
 import {GQLContext} from './graphql'
