+++ packages/server/postgres/types/index.d.ts
@@ -1,3 +1,6 @@
 import {Selectable} from 'kysely'
-import {OrganizationUser as OrganizationUserPG} from '../pg.d'
+import {OrganizationUser as OrganizationUserPG, TeamMember as TeamMemberPG} from '../pg.d'
+
 export type OrganizationUser = Selectable<OrganizationUserPG>
+
+export type TeamMember = Selectable<TeamMemberPG>
